api/requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
execnet==2.1.1
fonttools==4.53.1
idna==3.7
imageio==2.35.1
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe==2.1.5
networkx==3.3
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.1.2
pytest==8.3.2
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-rerunfailures==14.0
pytest-sugar==1.0.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
scipy==1.14.0
six==1.16.0
termcolor==2.4.0
tifffile==2024.8.10
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
